fix: x-slot no undef vars
