+++ lib/rules/valid-x-for.js
@@ -5,56 +5,12 @@
 'use strict';
 
 const jsxUtil = require('jsx-ast-utils');
-const eslintScope = require('eslint-scope');
 const { propName } = jsxUtil;
-const { isVariable } = require('../utils/tools');
+const { isVariable, addDeclaredVariables } = require('../utils/tools');
 const {
   DIR_NAME: { X_FOR },
   OPERATORS: { IN }
 } = require('../conf/directive-names');
-
-// # function helpers
-/**
- * add variables to current scope for 'x-for' directive
- * @param {object} context context
- * @param {Array} xForVars variables decclared by x-for directive
- */
-function addDeclaredVariables(context, xForVars) {
-  const globalScope = context.getScope();
-  xForVars.forEach(id => {
-    let variable = globalScope.set.get(id);
-
-    if (!variable) {
-      variable = new eslintScope.Variable(id, globalScope);
-      variable.eslintUsed = true;
-
-      globalScope.variables.push(variable);
-      globalScope.set.set(id, variable);
-    }
-  });
-
-  /*
-   * "through" contains all references which definitions cannot be found.
-   * So, here remove the variables from current "through" scope.
-   */
-  globalScope.through = globalScope.through.filter(reference => {
-    const name = reference.identifier.name;
-    const variable = globalScope.set.get(name);
-
-    if (variable) {
-      /*
-       * Links the variable and the reference.
-       * And this reference is removed from `Scope#through`.
-       */
-      reference.resolved = variable;
-      variable.references.push(reference);
-
-      return false;
-    }
-
-    return true;
-  });
-}
 
 //------------------------------------------------------------------------------
 // Rule Definition
