+++ lib/rules/valid-x-slot.js
@@ -11,7 +11,8 @@ const {
   isExactString,
   isXSlot,
   isPartlyXSlot,
-  lintMessage
+  lintMessage,
+  addDeclaredVariables
 } = require('../utils/tools');
 
 //------------------------------------------------------------------------------
@@ -34,7 +35,15 @@ module.exports = {
   },
 
   create: function(context) {
+    let xSlotVars = null;
     return {
+      'Program:exit'() {
+        console.log(xSlotVars)
+        if (xSlotVars) {
+          xSlotVars.length && addDeclaredVariables(context, xSlotVars);
+          xSlotVars = null;
+        }
+      },
       JSXOpeningElement(node) {
         node.attributes.forEach(attr => {
           if (attr.type !== 'JSXAttribute') return;
@@ -44,12 +53,20 @@ module.exports = {
           const isExactXSlot = isXSlot(tempName);
           if (isPartlyXSlot(tempName) && !isExactXSlot) {
             lintMessage(context, attr, 'validXSlotName');
-          } else if (
-            isExactXSlot &&
-            (tempValue !== true &&
-              (tempValue === '' || !isExactString(tempValue)))
-          ) {
-            lintMessage(context, attr, 'validXSlotValue');
+          } else if (isExactXSlot) {
+            if (
+              tempValue !== true &&
+              (tempValue === '' || !isExactString(tempValue))
+            ) {
+              lintMessage(context, attr, 'validXSlotValue');
+            } else if (tempValue && isExactString(tempValue)) {
+              xSlotVars = [
+                ...new Set([
+                  ...(xSlotVars || []),
+                  tempValue
+                ])
+              ];
+            }
           }
         });
       }
