+++ lib/utils/tools.js
@@ -6,6 +6,7 @@
 const {
   DIR_NAME: { X_SLOT }
 } = require('../conf/directive-names');
+const eslintScope = require('eslint-scope');
 const REG_X_SLOT = new RegExp(`^${X_SLOT}:[$A-Z_][0-9A-Z_$]*$`, 'i');
 const REG_PARTLY_X_SLOT = new RegExp(`^${X_SLOT}[:a-zA-Z$0-9_]*$`, 'i');
 
@@ -69,6 +70,49 @@ module.exports = {
       node,
       messageId,
       loc: node.loc
+    });
+  },
+
+  /**
+   * add variables to current scope for 'x-for/x-slot' directive
+   * @param {object} context context
+   * @param {Array} vars variables decclared by x-for/x-slot directive
+   */
+  addDeclaredVariables(context, vars) {
+    const globalScope = context.getScope();
+    vars.forEach(id => {
+      console.log(id);
+      let variable = globalScope.set.get(id);
+
+      if (!variable) {
+        variable = new eslintScope.Variable(id, globalScope);
+        variable.eslintUsed = true;
+
+        globalScope.variables.push(variable);
+        globalScope.set.set(id, variable);
+      }
+    });
+
+    /*
+    * "through" contains all references which definitions cannot be found.
+    * So, here remove the variables from current "through" scope.
+    */
+    globalScope.through = globalScope.through.filter(reference => {
+      const name = reference.identifier.name;
+      const variable = globalScope.set.get(name);
+
+      if (variable) {
+        /*
+        * Links the variable and the reference.
+        * And this reference is removed from `Scope#through`.
+        */
+        reference.resolved = variable;
+        variable.references.push(reference);
+
+        return false;
+      }
+
+      return true;
     });
   }
 };
